[io] Split RRawFile... in several compilation units
