+++ io/io/inc/ROOT/RRawFileDavix.hxx
@@ -1,0 +1,23 @@
+// @(#)root/io:$Id$
+// Author: Jakob Blomer
+
+/*************************************************************************
+ * Copyright (C) 1995-2018, Rene Brun and Fons Rademakers.               *
+ * All rights reserved.                                                  *
+ *                                                                       *
+ * For the licensing terms see $ROOTSYS/LICENSE.                         *
+ * For the list of contributors see $ROOTSYS/README/CREDITS.             *
+ *************************************************************************/
+
+#ifndef ROOT_RRawFileDavix
+#define ROOT_RRawFileDavix
+
+#include <ROOT/RRawFile.hxx>
+
+namespace ROOT {
+namespace Detail {
+
+} // namespace Detail
+} // namespace ROOT
+
+#endif
